Реализует модульность
